+++ client/pages/room.module.css
@@ -1,0 +1,166 @@
+.container {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100vh;
+  padding: 20px;
+  background-color: gray;
+}
+
+/* 左上の花火ピクセルアートの枠 */
+.artContainer {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 85%;
+  height: 90%;
+  border: 2px solid #333;
+}
+
+/* 花火のピクセルアート */
+.fireworks {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 300px;
+  height: 300px;
+  background-color: transparent;
+  transform: translate(-50%, -50%);
+}
+
+.fireworks::before,
+.fireworks::after {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  content: '';
+  background: conic-gradient(
+    from 0deg,
+    #f00 0deg 60deg,
+    #ff0 60deg 120deg,
+    #0f0 120deg 180deg,
+    #00f 180deg 240deg,
+    #f0f 240deg 300deg,
+    #f00 300deg 360deg
+  );
+  border-radius: 50%;
+  transform: scale(0); /* 初期状態で縮小 */
+  animation: explode 3s forwards ease-in-out; /* 1.5秒で1回実行 */
+}
+
+.fireworks::after {
+  transform: scale(0);
+  animation-delay: 0.75s; /* 少し遅れて二重のエフェクトを表現 */
+}
+
+@keyframes explode {
+  0% {
+    opacity: 0;
+    transform: scale(0); /* 爆発開始前は小さく */
+  }
+
+  50% {
+    opacity: 1;
+    transform: scale(1); /* 最大サイズに拡大 */
+  }
+
+  100% {
+    opacity: 0;
+    transform: scale(1.5); /* 徐々に消えていく */
+  }
+}
+
+/* 花火発射ボタンを枠の外の右下に配置 */
+.fireworkButton {
+  position: absolute;
+  right: -130px; /* 枠の外、右に50px移動 */
+  bottom: 0; /* 枠の外、下に50px移動 */
+  padding: 10px 20px;
+  font-size: 16px;
+  color: white;
+  cursor: pointer;
+  background-color: #09f5a2;
+  border: none;
+  border-radius: 5px;
+  transition: background-color 0.3s ease;
+}
+
+.fireworkButton:disabled {
+  cursor: not-allowed;
+  background-color: #ccc;
+}
+
+.fireworkButton:hover:not(:disabled) {
+  background-color: #09f5a2;
+}
+
+/* 吹き出しのスタイル */
+.chatBubbleContainer {
+  position: absolute;
+  top: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  justify-content: flex-start;
+  width: 100%;
+  padding-top: 20px;
+  padding-right: 20px;
+}
+
+.chatBubble {
+  max-width: 50%;
+  padding: 10px 20px;
+  margin: 10px 0;
+  color: white;
+  word-wrap: break-word;
+  background-color: #333;
+  box-shadow:
+    4px 0 black,
+    -4px 0 black,
+    0 4px black,
+    0 -4px black;
+  animation: slide-up 0.3s ease-in-out;
+}
+
+@keyframes slide-up {
+  from {
+    opacity: 0;
+    transform: translateY(10px);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+/* 下部のチャット選択肢 */
+.chatOptions {
+  position: fixed;
+  right: 20px;
+  bottom: 20px;
+  left: 20px;
+  display: flex;
+  justify-content: space-around;
+}
+
+.optionButton {
+  padding: 10px 20px;
+  color: #333;
+  cursor: pointer;
+  background-color: #fff;
+  border: 1px solid #ccc;
+  border-radius: 20px;
+  transition: background-color 0.3s ease;
+}
+
+.optionButton:hover {
+  background-color: #f0f0f0;
+}
+
+.logoImage {
+  width: 1100px;
+  height: 500px;
+}
